Cleanup code from product-related tool names
