+++ agents/agents-tools-registry/src/commonMain/kotlin/ai/grazie/code/agents/tools/registry/tools/ProjectTemplateTools.kt
@@ -6,7 +6,8 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
-object RiderProjectTemplateTools {
+object ProjectTemplateTools {
+
     abstract class CreateStructureTool : Tool<CreateStructureTool.Args, CreateStructureTool.Result>() {
         @Serializable
         data class Args(
